requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
et-xmlfile==1.1.0
idna==3.6
jmespath==1.0.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.2.0
pycparser==2.21
pypdfium2==4.27.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
tzdata==2024.1
urllib3==2.0.7
